extract units and icons
